+++ packages/common-server/package.json
@@ -45,6 +45,7 @@
     "analytics-node": "^5.1.0",
     "anymatch": "^3.1.2",
     "comment-json": "4.2.0",
+    "textextensions": "^5.15.0",
     "execa": "^5.0.0",
     "fs-extra": "^9.0.1",
     "gray-matter": "^4.0.2",
