+++ packages/plugin-core/package.json
@@ -1136,7 +1136,6 @@
     "replace-in-file": "^6.2.0",
     "semver": "^7.3.2",
     "string-similarity": "^4.0.4",
-    "textextensions": "^5.14.0",
     "unist-util-visit": "2.0.3",
     "vscode-languageclient": "^6.1.3"
   },
